Change counter type to size_t after a review comment

+++ silkworm/core/crypto/rmd160.c
@@ -368,8 +368,8 @@ void silkworm_rmd160(uint8_t out[20], const uint8_t* ptr, uint32_t len) {
     rmd160_init(buf);
 
     uint32_t current[16];
-    for (unsigned remaining = len; remaining >= 64; remaining -= 64) {
-        for (unsigned i = 0; i < 16; ++i) {
+    for (size_t remaining = len; remaining >= 64; remaining -= 64) {
+        for (size_t i = 0; i < 16; ++i) {
             current[i] = load32(ptr);
             ptr += 4;
         }
@@ -378,7 +378,7 @@ void silkworm_rmd160(uint8_t out[20], const uint8_t* ptr, uint32_t len) {
 
     rmd160_finish(buf, ptr, len);
 
-    for (unsigned i = 0; i < 20; i += 4) {
+    for (size_t i = 0; i < 20; i += 4) {
         out[i] = buf[i >> 2];
         out[i + 1] = buf[i >> 2] >> 8;
         out[i + 2] = buf[i >> 2] >> 16;
